requirements.txt
beautifulsoup4
openpyxl
pyyaml
requests